% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_scalars}
\alias{all_scalars}
\title{Check if all elements of a list are of length 1}
\usage{
all_scalars(x)
}
\arguments{
\item{x}{List.}
}
\value{
Logical flag, TRUE or FALSE.
}
\description{
Checks if lists contains onyl scalar values and can be safely unlisted,
maintaining the same length.
}
\keyword{internal}
