% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertion_messages}
\alias{assertion_messages}
\alias{.normal_message}
\alias{.bounds_message}
\title{Function closures for assertion failure messages}
\usage{
.normal_message(message)

.bounds_message(message)
}
\arguments{
\item{message}{a single string.}
}
\value{
A function \code{f} valid for \code{on_failure(fun) <- f} assignment.
}
\description{
Set of function closures made for easy assertion messages.
}
\keyword{internal}
