% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assertions}
\alias{assertions}
\alias{is_nullcount}
\alias{is_nullflag}
\title{Assertions}
\usage{
is_nullcount(x, limit = NULL)

is_nullflag(x)
}
\arguments{
\item{x}{Object to check.}

\item{limit}{Single positive integer, sets the upper bound of \code{x}. Does
nothing if NULL}
}
\value{
Logical vector of length one.
}
\description{
Check if \code{x} satisfies given conditions.
}
\section{Functions}{
\itemize{
\item \code{is_nullcount}: Checks if \code{x} is NULL or a single positive integer
<= \code{limit} (if limit is not \code{NULL}).

\item \code{is_nullflag}: Checks if \code{x} is a single logical, not-NA value or
NULL.
}}

\keyword{internal}
