% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgg_tools.R
\name{bgg_gameurl}
\alias{bgg_gameurl}
\title{Get BoardGameGeek URLs for games with given IDs}
\usage{
bgg_gameurl(ids)
}
\arguments{
\item{ids}{a numeric vector of positive integers.}
}
\value{
A character vector of the same length as \code{x}, contains URLs.
}
\description{
This function is a simple wrapper that returns URLs to given games by using
their IDs.
}
\examples{
bgg_gameurl(1:10)

}
\author{
Jakub Bujnowicz \email{bujnowiczgithub@gmail.com}
}
