% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggAnalytics-package.R
\docType{package}
\name{bggAnalytics-package}
\alias{bggAnalytics}
\alias{bggAnalytics-package}
\title{bggAnalytics: BoardGameGeek's Board Game Data Analysis Tools}
\description{
Tools for analysing board game data. Mainly focused on providing an interface for BoardGameGeek's XML API2 through R6 class system objects. More details about the BoardGameGeek's API can be obtained here \url{https://boardgamegeek.com/wiki/page/BGG_XML_API2}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JakubBujnowicz/bggAnalytics}
}

}
\keyword{internal}
