% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{.bgg_read_with_token}
\alias{.bgg_read_with_token}
\title{Get XML using an autherization token}
\usage{
.bgg_read_with_token(url, token)
}
\arguments{
\item{url}{a single string, an URL to a given website.}

\item{token}{a single string, the authorization token.}
}
\value{
A parsed XML.
}
\description{
Creates an XML from a given URL, using the provided BGG Authorization token.
}
\keyword{internal}
