% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{.bgg_readurl}
\alias{.bgg_readurl}
\title{Safely scrap HTML website}
\usage{
.bgg_readurl(url, ...)
}
\arguments{
\item{url}{a single string, an URL to a given website.}

\item{...}{other arguments passed to \code{read_html}.}
}
\value{
A parsed XML.
}
\description{
Opens the connection with \code{curl} and closes in case of error. Returns
more unambiguous error messages.
}
\keyword{internal}
