% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{.bgg_url}
\alias{.bgg_url}
\title{Return a given BoardGameGeek URL}
\usage{
.bgg_url(of)
}
\arguments{
\item{of}{Single string, either `api`, "ranking" or "boardgame".}
}
\value{
Single string with a page URL.
}
\description{
Get hyperlinks to given pages of BoardGameGeek site.
}
\keyword{internal}
