% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compress}
\alias{.compress}
\title{Compress vector into single string}
\usage{
.compress(vec, n_show = 5, collapse = ", ")
}
\arguments{
\item{vec}{Atomic vector.}

\item{n_show}{Positive integer.}

\item{collapse}{Single string, how to collapse the given vector.}
}
\value{
Single string.
}
\description{
Returns string which shows \code{n_show} first elements and \code{vec} and
amount of truncated elements.
}
\keyword{internal}
