% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_utils.R
\name{.extend_url_by_params}
\alias{.extend_url_by_params}
\title{Extend given URL by class params}
\usage{
.extend_url_by_params(url, params, class)
}
\arguments{
\item{url}{a single string.}

\item{params}{NULL or a list with params.}

\item{class}{a single string}
}
\value{
A single string with extended URL.
}
\description{
This takes an \code{url} and extends it by the given parameters for a
certain class
}
\keyword{internal}
