% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{.get_varspecs}
\alias{.get_varspecs}
\title{Get variable specification table for a given class}
\usage{
.get_varspecs()
}
\value{
The data.table with variable specifications
}
\description{
This can be used within \code{bggAPI} methods to extract rows for a proper
class. This also handles the 'pretty names'.
}
\keyword{internal}
