% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.internal_fun}
\alias{.internal_fun}
\title{Get one of the internal functions from the package}
\usage{
.internal_fun(fun_name)
}
\arguments{
\item{fun_name}{a single string.}
}
\value{
A function.
}
\description{
This makes sure internal functions from this package can be accessed which
sometimes fails when using \code{match.fun}.
}
\keyword{internal}
