% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.plural}
\alias{.plural}
\title{Add '-s' to the end of a string based on a count}
\usage{
.plural(string, count)
}
\arguments{
\item{string}{a single string.}

\item{count}{a single number that determines the addition of "-s".}
}
\value{
A single string.
}
\description{
Does exactly what the title says.
}
\keyword{internal}
