% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{.private_getter}
\alias{.private_getter}
\title{Getter for private slots of R6 objects}
\usage{
.private_getter(slotname)
}
\arguments{
\item{slotname}{Single string, name of private slot.}
}
\value{
Function that extracts the private `slotname` slot from a R6 object.
}
\description{
This function returns another function which is able to extract private
slots. It decreases redundant code and makes class definitions easier to
read.
}
\keyword{internal}
