% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_utils.R
\name{.process_params}
\alias{.process_params}
\title{Class parameters preprocessing}
\usage{
.process_params(params, class)
}
\arguments{
\item{params}{NULL or a list with params.}

\item{class}{a single string, class name.}
}
\value{
List with checked ans assigned parameters.
}
\description{
Process, validate and assign default values to parameters of a given class.
}
\keyword{internal}
