% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.split_acc2list}
\alias{.split_acc2list}
\title{Split according to a list}
\usage{
.split_acc2list(x, list)
}
\arguments{
\item{x}{an \code{R} object to split.}

\item{list}{a list.}
}
\value{
A list.
}
\description{
This splits \code{x} using \code{\link[base]{split}} by making sure that
the returned list is of the same length as \code{list} and every element
is of the same length as \code{list}'s elements.
}
\keyword{internal}
