% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{.switch_names}
\alias{.switch_names}
\title{Switch the style of object's data colnames}
\usage{
.switch_names(to)
}
\arguments{
\item{to}{a single string, style of names to switch to.}
}
\value{
Nothing, used for it's side effect.
}
\description{
This change between \code{PrettyName} and \code{Variable} names for variables
in \code{bggAPI} objects.
}
\keyword{internal}
