% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.unenclose}
\alias{.unenclose}
\title{Unenclose a closure}
\usage{
.unenclose(f)
}
\value{
A function with a modified environment.
}
\description{
Unenclose a closure by substituting names for values found in the
enclosing environment. Previously a part of \code{pryr} package.
}
\keyword{internal}
