% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{.xml_concatenate}
\alias{.xml_concatenate}
\title{Concatenate XML nodesets from a list}
\usage{
.xml_concatenate(xml_list)
}
\arguments{
\item{xml_list}{a list of XML nodesets.}
}
\value{
An XML nodeset.
}
\description{
Take a list of XML nodesets (e.g. outputs of \code{.xml_expand}) and combine
them together.
}
\keyword{internal}
