% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{.xml_expand}
\alias{.xml_expand}
\title{Get all 'items' nodes from XML Nodeset}
\usage{
.xml_expand(xml)
}
\arguments{
\item{xml}{an XML nodeset.}
}
\value{
XML nodeset.
}
\description{
This function expands \code{xml_nodeset} objects to all 'item' children. It
allows to use \code{lapply(., f)} per object (corresponding to one row in
\code{data}).
}
\keyword{internal}
