% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{expand_by}
\alias{expand_by}
\title{Generalised expand for every class}
\usage{
expand_by(variable_names = NULL, params = NULL)
}
\arguments{
\item{variable_names}{Character vector of variable names.}

\item{params}{List of parameters.}
}
\value{
Nothing, used for side effect.
}
\description{
It is a universal tool for fetching variables from XMLs and adding them
to \code{data} slot data.table.
}
\keyword{internal}
