% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{expand_xml}
\alias{expand_xml}
\title{Get all 'items' nodes from XML Nodeset}
\usage{
expand_xml(xml)
}
\arguments{
\item{xml}{XML Nodeset.}
}
\value{
XML Nodeset.
}
\description{
This function expands xml_nodeset to all 'item' children. It allows to use
`lapply(., f)` per Game ID.
}
\keyword{internal}
