% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{bggAnalytics options}
\description{
The following options are set by default by the package:
\itemize{
\item{\code{bggAnalytics.print}} [\code{10}] --- a single positive integer,
decides how many objects are printed.
\item{\code{bggAnalytics.verbose}} [\code{TRUE}] --- a logical value,
decides whether additional diagnostic information is printed while
fetching data.
\item{\code{bggAnalytics.username}} [\code{NULL}] --- a single string,
decides the default username value for \link{bggCollection} objects.
\item{\code{bggAnalytics.token}}[\code{NULL}] --- a single string, the BGG API
authorization token.
}
}
