% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_add}
\alias{path_add}
\alias{path_add_flag}
\alias{path_add_filter}
\title{Add params to APIs URL}
\usage{
path_add_flag(path, value)

path_add_filter(path, value)
}
\arguments{
\item{path}{String, URL to a given API.}

\item{value}{Variable with a value of parameter.}
}
\value{
String \code{path} expanded by \code{value}'s parameter.
}
\description{
Expand given path by a certain parameter. Values to value should be passed as
named variables, function uses internal substituting and deparsing. Correct
usage example is \code{test_param <- 1} and then
\code{path_add_filter("some_path", test_param)}. The result of that operation
is \code{"some_path&test_param=1"}.
}
\section{Functions}{
\itemize{
\item \code{path_add_flag}: Value should be logical or NULL. NULL and FALSE change nothing,
TRUE adds \code{&variable=1} at the end.

\item \code{path_add_filter}: Value should be a scalar (positive integer or logical)
or NULL. Logical values get casted to integer, so that path
is expanded by \code{&variable=0} or \code{&variable=1}. Integers
are pasted as is, i.e. for value of 5 we get \code{&variable=5}.
}}

\keyword{internal}
