% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{tokens}
\alias{tokens}
\alias{set_token}
\alias{get_token}
\title{BGG Authorization tokens}
\usage{
set_token(token)

get_token()
}
\arguments{
\item{token}{a single string, the authorization token.}
}
\description{
Required to handle most requests for BGG XML API2. For more information
please see "Using the XML API"'s section "Application Tokens".
}
\examples{
\dontrun{
    # Token taken from the example in the link above, just for reference
    set_token("e3f8c3ff-9926-4efc-863c-3b92acda4d32")

    get_token()
}

}
\references{
\href{https://boardgamegeek.com/using_the_xml_api}{Using the XML API}
}
