% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertions}
\alias{assertions}
\alias{.is_number}
\alias{.are_numbers}
\alias{.is_integer}
\alias{.are_integers}
\alias{.is_positive_integer}
\alias{.are_positive_integers}
\alias{.is_string}
\alias{.are_strings}
\alias{.is_boolean}
\alias{.are_booleans}
\alias{.is_nodeset}
\title{Assertions}
\usage{
.is_number(x, lb = NULL, ub = NULL)

.are_numbers(x, lb = NULL, ub = NULL)

.is_integer(x, lb = NULL, ub = NULL)

.are_integers(x, lb = NULL, ub = NULL)

.is_positive_integer(x, lb = NULL, ub = NULL)

.are_positive_integers(x, lb = NULL, ub = NULL)

.is_string(x, allowed = NULL)

.are_strings(x, allowed = NULL)

.is_boolean(x)

.are_booleans(x)

.is_nodeset(x)
}
\arguments{
\item{x}{an object to check.}

\item{lb, ub}{single numbers - (closed) bounds of \code{x}. Does
nothing if \code{NULL}.}
}
\value{
Logical vector of length one.
}
\description{
Check if \code{x} satisfies given conditions.
}
\keyword{internal}
