% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_tools.R
\name{assertions}
\alias{assertions}
\alias{.is_number}
\alias{.are_numbers}
\alias{.is_integer}
\alias{.are_integers}
\alias{.is_positive_integer}
\alias{.are_positive_integers}
\alias{.is_string}
\alias{.are_strings}
\alias{.is_boolean}
\alias{.are_booleans}
\title{Assertions}
\usage{
.is_number(x, limits = NULL)

.are_numbers(x)

.is_integer(x, limits = NULL)

.are_integers(x)

.is_positive_integer(x, limits = NULL)

.are_positive_integers(x, limits)

.is_string(x)

.are_strings(x)

.is_boolean(x)

.are_booleans(x)
}
\arguments{
\item{x}{Object to check.}

\item{limits}{Two numbs, sets the (closed) bounds of \code{x}. Does
nothing if \code{NULL}.}
}
\value{
Logical vector of length one.
}
\description{
Check if \code{x} satisfies given conditions.
}
\keyword{internal}
