% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggAnalytics-package.R
\docType{package}
\name{bggAnalytics-package}
\alias{bggAnalytics}
\alias{bggAnalytics-package}
\title{bggAnalytics: BGG's Board Game Data Analysis Tools}
\description{
Tools for analysing board game data. Mainly focused on providing 
    an interface for BGG's XML API2 through R6 class system objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JakubBujnowicz/bggAnalytics}
}

}
\keyword{internal}
