% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggAPI.R
\name{bggApi}
\alias{bggApi}
\alias{bggAPI}
\title{Main class for BGG XML API2}
\description{
Main class for BGG XML API2

Main class for BGG XML API2
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-switch_namestyle}{\code{bggAPI$switch_namestyle()}}
\item \href{#method-fetch}{\code{bggAPI$fetch()}}
\item \href{#method-expand}{\code{bggAPI$expand()}}
\item \href{#method-clone}{\code{bggAPI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-switch_namestyle"></a>}}
\if{latex}{\out{\hypertarget{method-switch_namestyle}{}}}
\subsection{Method \code{switch_namestyle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bggAPI$switch_namestyle(to)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetch"></a>}}
\if{latex}{\out{\hypertarget{method-fetch}{}}}
\subsection{Method \code{fetch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bggAPI$fetch(variable_names = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expand"></a>}}
\if{latex}{\out{\hypertarget{method-expand}{}}}
\subsection{Method \code{expand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bggAPI$expand(variable_names = NULL, params = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bggAPI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
