% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggGames.R
\name{bggGames}
\alias{bggGames}
\title{Generalised Distribution Object}
\description{
A generalised distribution object for defining custom probability distributions
  as well as serving as the parent class to specific, familiar distributions. Common
  mathematical and statistical methods for distributions are defined here with approximate numerical
  calculations (as opposed to analytical results).
}
