% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_utils.R
\name{extraction_functions}
\alias{extraction_functions}
\alias{.nodes2text}
\alias{.nodes2number}
\alias{.nodes2logical}
\alias{.attr2text}
\alias{.attr2number}
\alias{.attr2logical}
\title{Extraction functions}
\usage{
.nodes2text(xml, xpath)

.nodes2number(xml, xpath)

.nodes2logical(xml, xpath)

.attr2text(xml, xpath, attr)

.attr2number(xml, xpath, attr)

.attr2logical(xml, xpath, attr)
}
\arguments{
\item{xml}{XML item list.}

\item{xpath}{Single string, name of XPATH expression, for example node name.}

\item{attr}{Single string, name of attribute}
}
\value{
List of elements of equal length as \code{xml}.
}
\description{
These functions extract data from given nodes/attributes from XML item list.
}
\keyword{internal}
